users/pyproject.toml
[tool.poetry]
name = "ffbc-auth"
version = "0.1.0"
description = ""
authors = ["Akim Abdulaev <mr.world008@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
fastapi = "^0.75.0"
python-multipart = "^0.0.5"
email-validator = "^1.1.3"
SQLAlchemy = "^1.4.32"
psycopg2-binary = "^2.9.3"
pytest = "^7.0.1"
celery = "^5.2.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.0.1"
emails = "^0.6"
gunicorn = "^20.1.0"
alembic = "^1.7.6"
uvicorn = "^0.17.5"
python-dotenv = "^0.19.2"
databases = {extras = ["asyncpg"], version = "^0.5.5"}
asyncpg = "^0.25.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = "^4.3.4"
fastapi-jwt-auth = "^0.5.0"
pilkit = "^2.0"
google-cloud-storage = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
